+++ src/services/index.ts
@@ -32,19 +32,4 @@ export {
   resolveClickFilters,
 } from "./admin-management";
 
-export {
-  createBundle,
-  listBundles,
-  getBundle,
-  updateBundle,
-  archiveBundle,
-  unarchiveBundle,
-  deleteBundle,
-  addLinkToBundle,
-  removeLinkFromBundle,
-  getBundleAnalytics,
-  listBundleLinks,
-  listBundlesForLink,
-} from "./bundle-management";
-
 export type { ServiceResult } from "./result";